refactor: 디렉터리명과 중복되는 Card prefix를 제거
- 불필요하게 네이밍이 길어진 것을 수정
