+++ src/components/paymentInputPage/cardInputForm/cardInput/CVCInput.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { expect, userEvent, within } from '@storybook/test';
 import styles from '../../../common/inputForm/input/Input.module.css';
-import CardCVCInput from './CardCVCInput';
+import CardCVCInput from './CVCInput';
 
 const meta = {
   title: 'CardCVCInput',
